+++ packages/rocketh-export/src/index.ts
@@ -57,7 +57,7 @@ export async function run(
 			address: d.address,
 			linkedData: d.linkedData,
 			bytecode: options.includeBytecode ? d.bytecode : undefined,
-			startBlock: d.receipt?.blockNumber ? parseInt(d.receipt.blockNumber) : undefined,
+			startBlock: d.receipt?.blockNumber ? parseInt(d.receipt.blockNumber.slice(2), 16) : undefined,
 		})),
 		name: config.networkName,
 	};