+++ packages/hardhat-rocketh/CHANGELOG.md
@@ -1,5 +1,12 @@
 # hardhat-rocketh
 
+## 0.6.23
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.6.15
+
 ## 0.6.22
 
 ### Patch Changes